display inline in-app WebView within sample app
Part of: https://linear.app/customerio/issue/MBL-295/[spike]-determine-amount-of-effort-required-for-react-native-wrapper

We are able to display the native iOS UIKit View within a RN sample app. This commit shows the code required to do this. This commit gives us an idea of the effort required to create a wrapper around a native iOS View and display it in an app. Also, it shows WebViews in RN behave similarly as in a native iOS app.