get apn app working again with weird node issue
got error "Can't find the 'node' binary to build the React Native bundle." when trying to compile. I use asdf on my computer for installing node which is probably why I get this error.

I created a new RN app and viewed the xcode project build phases that new apps use. It's identical to the FCM sample app. So I edited the Build Phase that threw this error with the new build phase content to fix this error.